[Refactor][Ops] Read dtype at forward for the reduction and norm families
20 op constructors stop taking `dtype`; each reads it from its input in
`forward()` and keys its kernel cache by shape and dtype.

Also removes the `_committed_dtype` dual path (an optional ctor dtype,
validated when supplied) that cumulative, softmax and the four
ada/fused-add norm ops carried. `LayerNormFwdOp` gains a real kernel
cache — it kept one kernel and rebuilt whenever the leading-dims product
changed, which with a second dtype in play would rebuild every call.

`_dtype_codegen` now emits the validator body unrolled per input instead
of looping over `input_names` through `locals()`. Dynamo cannot trace
`locals()`, and this body runs inside `forward()`, so an op that
validates dtypes would lose `fullgraph`. The unreachable
`same_as(ref)`-not-supplied branch goes with it: synthesis already
rejects a ref that names no sibling input.

`_SoftmaxBaseOp` keeps its inline dtype-union check rather than
delegating to `_validate_dtypes`: `LogSumExpFwdOp`'s manifest entry
declares only float16|bfloat16 while the kernel and its tests use
float32. Narrowing the code to match is the wrong direction, and
widening the manifest is a separate change.

Two tests asserted that a ctor dtype disagreeing with the input raises.
That coupling is gone, but the invariant underneath is not — the `dim=[]`
short-circuit must still gate dtype — so they now feed a dtype outside
the manifest union instead of one outside the ctor's.

+++ benchmarks/ops/bench_ada_layer_norm.py
@@ -29,7 +29,7 @@ def test_ada_layer_norm_bench(m: int, n: int, dtype: torch.dtype) -> None:
     test = AdaLayerNormWorkload(m, n, dtype)
     inputs = test.gen_inputs()
 
-    op = AdaLayerNormFwdOp(dtype=dtype)
+    op = AdaLayerNormFwdOp()
     bm = ManifestBenchmark(_ADA_OP_NAME, op, test)
     result = bm.profile(op, *inputs)
     BenchmarkReport.record(op, locals(), result, tag="tileops")
@@ -48,7 +48,7 @@ def test_ada_layer_norm_zero_bench(m: int, n: int, dtype: torch.dtype) -> None:
     test = AdaLayerNormZeroWorkload(m, n, dtype)
     inputs = test.gen_inputs()
 
-    op = AdaLayerNormZeroFwdOp(dtype=dtype)
+    op = AdaLayerNormZeroFwdOp()
     bm = ManifestBenchmark(_ADA_ZERO_OP_NAME, op, test)
     result = bm.profile(op, *inputs)
     BenchmarkReport.record(op, locals(), result, tag="tileops")
